+++ src/lib/Scenes/Artwork/Components/ImageCarousel/__tests__/ImageWithLoadingState-tests.tsx
@@ -1,15 +1,15 @@
 import { mount } from "enzyme"
+import OpaqueImageView from "lib/Components/OpaqueImageView"
 import React from "react"
-import { Image } from "react-native"
 import { ImageWithLoadingState } from "../ImageWithLoadingState"
 
-const source = { uri: "https://image.com/image.jpg" }
+const imageURL = "https://image.com/image.jpg"
 const style = { width: 100, height: 300 }
 
 describe(ImageWithLoadingState, () => {
   it("renders the image", () => {
-    const wrapper = mount(<ImageWithLoadingState source={source} style={style} />)
-    expect(wrapper.find(Image)).toHaveLength(1)
-    expect(wrapper.find(Image).props().source).toBe(source)
+    const wrapper = mount(<ImageWithLoadingState imageURL={imageURL} {...style} />)
+    expect(wrapper.find(OpaqueImageView)).toHaveLength(1)
+    expect(wrapper.find(OpaqueImageView).props().imageURL).toBe(imageURL)
   })
 })