+++ CHANGELOG.md
@@ -19,6 +19,9 @@
 ### 1.12.4
 
 - Add shim for calling view in room view.
+- Fix carousel alignment - ds300
+- Fix styled-components warning - ds300
+- Use OpaqueImageView in ImageCarousel -ds300
 
 ### 1.12.3